Solution to Sort Students by Marks

+++ Sort Students by Marks@@67877d61dc107486ad5c2748/code.c
@@ -20,7 +20,7 @@ int main(){
     int n;
     scanf("%d",&n);
     struct Student std[n];
-    int arr[n];
+    float arr[n];
     for(int i=0;i<n;i++){
         scanf("%d",&std[i].roll);
         scanf("%s",std[i].name);
